+++ src/days/Day3.java
@@ -14,7 +14,7 @@ public class Day3 extends Day {
     }
 
     @Override
-    public int doPart1() {
+    public long doPart1() {
         int sum = 0;
 
         for (ArrayList<Integer> num : numsInfo) {
@@ -30,7 +30,7 @@ public class Day3 extends Day {
     }
 
     @Override
-    public int doPart2() {
+    public long doPart2() {
         int sum = 0;
 
         for (ArrayList<Integer> sym : symbolsInfo) {
